docs: Don't suggest configuring hk in config env (#424)
I was trying to disable staging, and since config `stage` doesn't exist,
tried to configure via `env`, but that didn't work either.

I poked through `src/settings.rs` and didn't see anything obvious that
suggests this is supposed to work.

So, removing the misleading section from the docs (alternatively we
could make this supported, but this doesn't preclude that)