feat: add check-conventional-commit utility and built-in (#477)
This implements a simple util for checking whether the commit title
(i.e. first line of the commit) adheres to conventional commits.

Admittedly this is a fairly naive implementation, but I thought it might
serve as a nice starting point for folks who want to enforce
conventional commits.

I chose to use default set of commit types from
[`@commitlint/config-conventional`](https://github.com/conventional-changelog/commitlint/tree/master/%40commitlint/config-conventional)
as the baseline, and provided a flag for overriding them with a
comma-separated list.

EDIT: need to add docs, etc. I'll mark the PR ready once I've done that.

---------

Signed-off-by: Joonas Bergius <joonas@bergi.us>